validate jobs in Job.find
In the case of two or more Sidekiq workers the job hash in Redis can be
left in a partly deleted state after one worker deletes the job and
another worker sets the last_enqueue_time value. If this happens
before this commit then calls to Job.find will return an invalid job
that doesn't have a  name (or any properties but last_enqueue_time)
and can't be destroyed. After this commit nil is returned if the Job is
not valid.