Only check out a Redis connection if necessary (#438)
In Sidekiq 7, calls to `Sidekiq.redis` checks out a connection from an
internal connection pool instead of the connection pool used by the
job processors. If a connection is checked out but not used, it's
possible that the heartbeat thread won't keep the connection alive
before the Redis server client timeout.

To avoid this, update the checks to return earlier if the Redis
connection is not needed.

This might help reduce intermittent `Broken pipe` errors as reported
in https://github.com/redis-rb/redis-client/issues/119.