Readme: clarify that job status gets persisted
It was not obvious from the docs that this field got persisted when
the job reloaded if it wasn't set explicitly. That means that a job
that is disabled in the web UI would still be disabled after a
redeployment. This just adds some more documentation to make that
abundantly clear.