Fix usage of ActiveJob::Base.queue_name
In `activejob` the class method `MyJob.queue_name` returns a `Proc`.
To obtain the actual queue name, we need to invoke the method on
an instance of the job class, e.g. `MyJob.new.queue_name`.

The test suite was unable to detect this problem because a mocked
version of the `ActiveJob::Base` class was used which did not act
in the same way as the actual `activejob` gem.

We fix this by testing against the actual `activejob` gem.

Fixes #512.